refactor: move secret output handler classes to secret.output
