--- ggshield/output/gitlab_webui/__init__.py
@@ -1,4 +1,0 @@
-from .gitlab_webui_output_handler import GitLabWebUIOutputHandler
-
-
-__all__ = ["GitLabWebUIOutputHandler"]
