--- ggshield/output/json/__init__.py
@@ -1,4 +1,0 @@
-from .json_output_handler import JSONOutputHandler
-
-
-__all__ = ["JSONOutputHandler"]
