--- ggshield/output/text/__init__.py
@@ -1,4 +1,0 @@
-from .text_output_handler import TextOutputHandler
-
-
-__all__ = ["TextOutputHandler"]
