+++ ggshield/secret/output/__init__.py
@@ -1,0 +1,12 @@
+from .gitlab_webui_output_handler import GitLabWebUIOutputHandler
+from .json_output_handler import JSONOutputHandler
+from .output_handler import OutputHandler
+from .text_output_handler import TextOutputHandler
+
+
+__all__ = [
+    "OutputHandler",
+    "JSONOutputHandler",
+    "TextOutputHandler",
+    "GitLabWebUIOutputHandler",
+]
