+++ ggshield/secret/output/output_handler.py
@@ -51,8 +51,6 @@ class OutputHandler(ABC):
         raise NotImplementedError()
 
     def _get_exit_code(self, scan: ScanCollection) -> ExitCode:
-        if scan.has_iac_result:
-            return ExitCode.SCAN_FOUND_PROBLEMS
         if self.ignore_known_secrets:
             if scan.has_new_secrets:
                 return ExitCode.SCAN_FOUND_PROBLEMS
