+++ ggshield/secret/output/text_output_handler.py
@@ -8,10 +8,7 @@ from ggshield.core.errors import UnexpectedError
 from ggshield.core.filter import censor_content, leak_dictionary_by_ignore_sha
 from ggshield.core.text_utils import Line, pluralize
 from ggshield.core.utils import Filemode, find_match_indices, get_lines_from_content
-from ggshield.output.output_handler import OutputHandler
-from ggshield.scan import Result, ScanCollection
-
-from .message import (
+from ggshield.output.text.message import (
     file_info,
     flatten_policy_breaks_by_line,
     leak_message_located,
@@ -20,7 +17,10 @@ from .message import (
     policy_break_header,
     secrets_engine_version,
 )
-from .utils import get_offset, get_padding
+from ggshield.output.text.utils import get_offset, get_padding
+from ggshield.scan import Result, ScanCollection
+
+from .output_handler import OutputHandler
 
 
 class TextOutputHandler(OutputHandler):
