+++ tests/unit/output/test_gitlab_webui_output.py
@@ -1,6 +1,6 @@
 from pygitguardian.models import Match, PolicyBreak
 
-from ggshield.output.gitlab_webui.gitlab_webui_output_handler import format_policy_break
+from ggshield.secret.output.gitlab_webui_output_handler import format_policy_break
 
 
 def test_format_policy_break():
